Add a percentage of max parameter to registration
Basically provides a parameter to determine the lower bound required to
be met for a particular shift to be deemed a match. For example, `1.0`
would mean that the shift would need to generate the maximum overlap.
Alternatively `0.9` would mean that the shift would need to reach 90% of
the maximum overlap or better to be included. As we have seen shifts to
be somewhat erratic from frame to frame (resulting in very strange
results), this tuning parameter should provide a way to relax the
requirements expected of a frame shift.