Drop usage of `set_num_workers` in the workflow
We no longer have a `CORES` environment variable that we need to set so
one of `set_num_workers` use cases is unneeded. Further following the
user's specifications on the number of workers to user is more
straightforward now. In other words, we don't need to factor in a
scheduler process. Given these changes just use `psutil`'s `cpu_count`
directly in the workflow if `num_workers` is not provided. This doesn't
handle the `-1` case. However it isn't that big of a deal for the user
to adapt as this is user facing code that is deployed with the user for
analysis.