Persist denoised images before shifting
As computing the shifted denoised images means computing the minimum of
the denoised images, it makes sense to persist the denoised images
before finding the minimum. Otherwise this filtering process may need to
be performed twice. Once to get the minimum and a second time to
subtract it out. Thus persisting before computing the minimum guarantees
this computation need only be performed once.