Bump the test runtime limit to 180
Seems we are now going over the runtime limit for the Notebook executor.
Not totally sure why, but it doesn't seem that we are going that far
over as the tests do sometimes complete within the limit. It may be a
dependency related change. Go ahead and bump the timeout another 60secs.
Should afford the Notebook tests enough time to complete.