Provide option to use Zopfli
Includes an option to use Zopfli to perform gzip compression. This can
make a small though notable improvement in compression size of the data.
Though can take a bit longer to run. As the data is still gzip
compressed, it can still be decompressed with any gzip implementation
like Pako. So no other changes are needed to handle the data.