Drop initial inner stats from dictionary learning
As we are now able to defer to scikit-learn to initialize the inner
statistics for the online dictionary learning algorithm, drop our
initialization step for the inner statistics. This should simplify the
Dask graph a little bit and inline the allocation of these arrays to
within scikit-learn itself. Also should avoid having our inner stats
initialization code here go stale relative to what scikit-learn includes
for this step.